Refactor SolarCLIPImageEmbedder module and add projection options
This commit refactors the SolarCLIPImageEmbedder module in the encoders/modules.py file. It introduces the ability to choose different projection types for the image embeddings, including "ConvTrans" and "Linear". The "ConvTrans" projection uses a series of transposed convolutional layers, while the "Linear" projection uses a linear layer followed by reshaping. This update allows for more flexibility in generating image embeddings.